Fixed implicit typecasts

+++ Sources/Extras/CoreMediaIO/DeviceAbstractionLayer/Devices/Sample/Assistant/Server/Device/CMIO_DPA_Sample_Server_Frame.cpp
@@ -68,7 +68,7 @@ namespace CMIO { namespace DPA { namespace Sample { namespace Server
 		mDroppedFrameCount(droppedFrameCount),
 		mFirstFrameTime(firstFrameTime),
 		mBufferID(bufferID),
-		mSize(size),
+		mSize(static_cast<UInt32>(size)),
 		mFrameData(data),
 		mClients(),
 		mClientsMutex("frame clients mutex")
